// evcar-news/admin/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CurrencyDto: {
            /** @example EGP */
            code: string;
            /** @example جنيه مصري */
            nameAr: string;
            /** @example Egyptian Pound */
            nameEn: string;
            /** @example ج.م */
            symbolAr: string | null;
            /** @example E£ */
            symbolEn: string | null;
            /** @example 2 */
            decimals: number;
        };
// ...
        CreateCurrencyDto: {
            /**
             * @description ISO-4217 code.
             * @example KWD
             */
            code: string;
            /** @example دينار كويتي */
            nameAr: string;
            /** @example Kuwaiti Dinar */
            nameEn: string;
            /** @example د.ك */
            symbolAr?: string | null;
            /** @example KD */
            symbolEn?: string | null;
            /** @default 2 */
            decimals: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}